add 15 new code

+++ exercises/conversions/using_as.rs
@@ -10,11 +10,10 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
@@ -31,3 +30,24 @@ mod tests {
         assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
     }
 }
+
+
+// fn average(values: &[f64]) -> f64 {
+//     let total = values.iter().sum::<f64>();
+//     total / values.len()
+// }
+
+// fn main() {
+//     let values = [3.5, 0.3, 13.0, 11.7];
+//     println!("{}", average(&values));
+// }
+
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
+
+//     #[test]
+//     fn returns_proper_type_and_value() {
+//         assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
+//     }
+// }
